[HierarchicalGenePool]: Fixed getFittest() method
Fixed the getFittest() method in the HierarchicalGenePool class so it
actually returns the fittest individual. This was a holdover from when
the population was guaranteed to be sorted - when this was the case,
returning the fittest individual meant simply pulling the first one from
the list - design changes made since then mean the population is no longer
guaranteed to be sorted.

Signed-off-by: Sebastian Lenartowicz <slenarto@uoguelph.ca>